# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Manager Box API",
        "description": "Documentação da API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desenvolvimento"
        }
    ],
    "paths": {
...
        "/refresh-token":{
            "post": {
                "summary": "Revalidate the user access token",
                "tags": ["Account"],
                "operationId": "refreshToken",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {

                                    "refresh_token": {
                                        "description": "User refresh token",
                                        "required": true,
                                        "type": "string",
                                        "example": "HUSy4zYcxUalIejfvE7s1ZuTLrNIsU1pr9Xucf4ouaMAp4eKSRSceHmT0AFli9Vm",
                                        "minLength": 64,
                                        "maxLength": 64,
                                        "pattern": "\\D{64}"
                                    }

                                }
                            }
                        }
                    }
                },
                "responses": {

                    "200": {
                        "description": "Token revalidated successfully",
                        "headers": {
                            "Set-Cookie": {
                                "description": "User refresh token",
                                "schema": {
                                    "type": "string",
                                    "example": "refresh_token=HUSy4zYcxUalIejfvE7s1ZuTLrNIsU1pr9Xucf4ouaMAp4eKSRSceHmT0AFli9Vm; expires=Wed, 12 Feb 2025 00:16:13 GMT; Max-Age=604800; path=/; domain=1; secure; httponly; samesite=lax"
                                }
                            }
                        },
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {

                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },

                                        "access_token": {
                                            "type": "string",
                                            "example": "1|j5uIgnf39o3V7yiSBg97QgHYf0zXEuI1O42nw8mcIN6h4DSpJzu4sr8NKeTwwgkB"
                                        }

                                    }
                                }
                            }
                        }
                    },

                    "400": {
                        "description": "User has sent a bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {

                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },

                                        "message": {
                                            "type": "string",
                                            "example": "Bad Request"
                                        },

                                        "errors": {
                                            "type": "object",
                                            "properties": {
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "refresh_token is a required field"
                                                }
                                            }
                                        }

                                    }
                                }
                            }
                        }
                    },

                    "401": {
                        "description": "User has sent a invalid or expired token",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {

                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },

                                        "errors": {
                                            "type": "object",
                                            "properties": {
                                                "refresh_token": {
                                                    "type": "string",
                                                    "example": "Invalid or expired refresh token"
                                                }
                                            }
                                        }

                                    }
                                }
                            }
                        }
                    },

                    "500": {
                        "description": "An internal server error has occurred",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "$ref": "#/components/schemas/InternalServerErrorResponse"
                                }
                            }
                        }
                    }

                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
            "InternalServerErrorResponse": {
                "type": "object",
                "properties": {

                    "success": {
                        "type": "boolean",
                        "example": false
                    },

                    "message": {
                        "type": "string",
                        "example": "Internal Server Error"
                    }

                }
            },
...
        }
    }
}